+++ spell.py
@@ -1,6 +1,6 @@
 class Spell:
 
-	def __init__(self, name: str = 'Spell', damage: float = 0, mana_cost: float = 0, cast_range: float = 0):
+	def __init__(self, name: str = 'Spell', damage: int = 0, mana_cost: int = 0, cast_range: int = 0):
 		self.name = name
 		self.damage = damage
 		self.mana_cost = mana_cost
