+++ weapon.py
@@ -1,5 +1,5 @@
 class Weapon:
 
-	def __init__(self, name: str = 'Weapon', damage: float = 0):
+	def __init__(self, name: str = 'Weapon', damage: int = 0):
 		self.name = name
 		self.damage = damage
